Fix MarkedPointProcess fitting

+++ test/marked_poisson_process.jl
@@ -12,8 +12,8 @@ pp = MarkedPoissonProcess(1.0, Categorical([0.1, 0.3, 0.6]))
 h1 = rand(rng, pp, 0.0, 1000.0)
 h2 = simulate_ogata(rng, pp, 0.0, 1000.0)
 
-pp_est1 = fit(MarkedPoissonProcess{Float32,Int,Categorical}, [h1, h1])
-pp_est2 = fit(MarkedPoissonProcess{Float32,Int,Categorical}, [h2, h2])
+pp_est1 = fit(MarkedPoissonProcess{Int,Float32,Categorical}, [h1, h1])
+pp_est2 = fit(MarkedPoissonProcess{Int,Float32,Categorical}, [h2, h2])
 
 λ_error1 = mean(abs, pp_est1.λ - pp.λ)
 λ_error2 = mean(abs, pp_est2.λ - pp.λ)
